refactor: move unnecessary dependencies to devDependencies (#444)
The dev-interface had angular dependencies with a specific version listed in its dependencies. In a (npm/yarn) workspace containing another angular version this could cause runtime conflicts as a different version of a angular module could be loaded (Depending on how dependencies are resolved. 

As the dev-interface is already shipping an app build (the dist directory), angular doesn't have to be listed as a dependency as it is a build-time only depdency.